kdtree: simplify KdTreeV construction (unsorted Node data)
The search algorithm does not depend on sorted data in the nodes. It
suffices to execute min or max reduction on the data and determine its
index.

This simplifies Tree construction (Node::insert) as the shifting of
parts of SIMD vectors becomes unnecessary. Instead one Point needs to be
extracted and in its place the new Point must be inserted. This reduces
the code to simple scalar load and stores to memory. Also the code
becomes much more concise.

Other changes:
* replacing T in main() with an integer type is now easily possible
because of the change for the uniform distribution type.
* flush cout after writing the line on insert performance
* get_kdtree_value is more flexible: non-const overload and
non-constexpr index overloads

Signed-off-by: Matthias Kretz <kretz@kde.org>