kdtree: less Scalar -> simdized conversions, no distance recalculation
* Node::findNearest basically only requires the search point as
broadcast to simdize<T>. Thus have KdTreeV::findNearest do the broadcast
once and then pass a reference to it to all subsequent Node::findNearest
functions.
* Node::findNearest determines the best candidate in its subtree, but it
already new the distance to the search point for that candidate. Before
this distance was then recalculated in the calling function (scalar
variant). Now Node::findNearest simply returns a std::pair with the
candidate point and the distance. This reduces the number of distance
calculations considerably (and they are all vectorized now).

Signed-off-by: Matthias Kretz <kretz@kde.org>